+++ src/paired_anc.h
@@ -1,2 +1,2 @@
 int est_paired_anc_gl(int nSites, int K, int nKs, double *gl1, double **f, double *res2);
-int est_paired_anc_gt(int nSites, int K, int nKs, int *gt1, double **f, double *res2);
+int est_paired_anc_gt(int nSites, int K, int nKs, unsigned short int *gt1, double **f, double *res2);
